+++ Sources/SwiftTUIViews/Collections/Table.swift
@@ -104,7 +104,10 @@ extension Table {
     in context: ResolveContext
   ) -> ResolvedNode {
     let styleEnvironment = context.environmentValues.styleEnvironmentSnapshot
-    let isFocused = context.environmentValues.focusedIdentity == context.identity
+    let isFocused =
+      context.environmentValues.focusedIdentity(
+        comparedAgainst: [context.identity]
+      ) == context.identity
     let isEnabled = context.environmentValues.isEnabled
     let showsFocusEffect = context.environmentValues.isFocusEffectEnabled
     let isSelectable = selectionPolicy.isSelectable